Move to_container tests from test_base_config.py to test_to_container.py
